Vinna apps can now haz sessions.
This implementation is pretty strait forward in that there now interface/implementation. This could be changed later if the need arises (e.g. for tests).

How it works:
Upon receiving a request, VinnaFilter checks if there's an existing servlet session and that a custom key exists. If that's the case, a vinna.Session instance is retrieved using that key. Otherwise, Vinna.newSession() is called to create a new instance. Either way, that instance is placed in VinnaContext.

Therefore, there's always a vinna.Session instance, regardless of the existence of a backing servlet session. vinna.Session exposes a couple of methods:
- create: to create a backing servlet session. This should be called if you want to be able to store values in it, otherwise calls to put will throw an exception
- get/put: a Map like low-level API to get/set values in the session.
- delete: to delete the backing servlet session
- renew: to renew the backing session while keeping the values it contained. Could be useful to mitigate session-fixation attacks.

After executing the yielded response (if the route matches), the filter checks again if there's a backing http session. If none is found, then there's nothing left to do: either there wasn't a session to start with, or the app decided to delete it. Otherwise, the vinna.Session instance created earlier is placed in the servlet session under a custom key.

Storing the vinna.Session instance as a whole in the servlet session is not strictly required to get/set values in the session. It is done to enable the apps to use type-safe session APIs. A user app can override Vinna.newSession() to return a subclass with fields, getters and setters. The controllers could then cast the result of VinnaContext.get().session (typed vinna.Session) to the subclass type and use the getters/setters instead of String keys and raw Object values. And since the whole vinna session instance is persisted, the fields it contains would get persisted too. Credits go to Wicket for this idea.

Closes #16.